Extract edit fade curve painting
